BAU: Move client session id to common test variables
This cleans up test setup and will also enable us to introduce shared common headers to tidy up the tests
